refactor(auth): rename Form -> SignForm
